third_party/blink/web_tests/accessibility/textarea-caret-position-expected.txt

This tests that caret position is reported correctly for textarea elements.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS textareaAccessible is textareaAccessible.selectionAnchorObject
PASS textareaAccessible.selectionAnchorOffset is 0
PASS textareaAccessible is textareaAccessible.selectionFocusObject
PASS textareaAccessible.selectionFocusOffset is 0
PASS textareaAccessible is textareaAccessible.selectionAnchorObject
PASS textareaAccessible.selectionAnchorOffset is 1
PASS textareaAccessible is textareaAccessible.selectionFocusObject
PASS textareaAccessible.selectionFocusOffset is 1
PASS textareaAccessible is textareaAccessible.selectionAnchorObject
PASS textareaAccessible.selectionAnchorOffset is 2
PASS textareaAccessible is textareaAccessible.selectionFocusObject
PASS textareaAccessible.selectionFocusOffset is 2
PASS textareaAccessible is textareaAccessible.selectionAnchorObject
PASS textareaAccessible.selectionAnchorOffset is 3
PASS textareaAccessible is textareaAccessible.selectionFocusObject
PASS textareaAccessible.selectionFocusOffset is 3
PASS textareaAccessible is textareaAccessible.selectionAnchorObject
PASS textareaAccessible.selectionAnchorOffset is 4
PASS textareaAccessible is textareaAccessible.selectionFocusObject
PASS textareaAccessible.selectionFocusOffset is 4
PASS textareaAccessible is textareaAccessible.selectionAnchorObject
PASS textareaAccessible.selectionAnchorOffset is 5
PASS textareaAccessible is textareaAccessible.selectionFocusObject
PASS textareaAccessible.selectionFocusOffset is 5
PASS textareaAccessible is textareaAccessible.selectionAnchorObject
PASS textareaAccessible.selectionAnchorOffset is 6
PASS textareaAccessible is textareaAccessible.selectionFocusObject
PASS textareaAccessible.selectionFocusOffset is 6
PASS textareaAccessible is textareaAccessible.selectionAnchorObject
PASS textareaAccessible.selectionAnchorOffset is 7
PASS textareaAccessible is textareaAccessible.selectionFocusObject
PASS textareaAccessible.selectionFocusOffset is 7
PASS textareaAccessible is textareaAccessible.selectionAnchorObject
PASS textareaAccessible.selectionAnchorOffset is 8
PASS textareaAccessible is textareaAccessible.selectionFocusObject
PASS textareaAccessible.selectionFocusOffset is 8
PASS textareaAccessible is textareaAccessible.selectionAnchorObject
PASS textareaAccessible.selectionAnchorOffset is 9
PASS textareaAccessible is textareaAccessible.selectionFocusObject
PASS textareaAccessible.selectionFocusOffset is 9
PASS textareaAccessible is textareaAccessible.selectionAnchorObject
PASS textareaAccessible.selectionAnchorOffset is 10
PASS textareaAccessible is textareaAccessible.selectionFocusObject
PASS textareaAccessible.selectionFocusOffset is 10
PASS textareaAccessible is textareaAccessible.selectionAnchorObject
PASS textareaAccessible.selectionAnchorOffset is 11
PASS textareaAccessible is textareaAccessible.selectionFocusObject
PASS textareaAccessible.selectionFocusOffset is 11
PASS textareaAccessible is textareaAccessible.selectionAnchorObject
PASS textareaAccessible.selectionAnchorOffset is 12
PASS textareaAccessible is textareaAccessible.selectionFocusObject
PASS textareaAccessible.selectionFocusOffset is 12
PASS textareaAccessible is textareaAccessible.selectionAnchorObject
PASS textareaAccessible.selectionAnchorOffset is 13
PASS textareaAccessible is textareaAccessible.selectionFocusObject
PASS textareaAccessible.selectionFocusOffset is 13
PASS textareaAccessible is textareaAccessible.selectionAnchorObject
PASS textareaAccessible.selectionAnchorOffset is 14
PASS textareaAccessible is textareaAccessible.selectionFocusObject
PASS textareaAccessible.selectionFocusOffset is 14
PASS textareaAccessible is textareaAccessible.selectionAnchorObject
PASS textareaAccessible.selectionAnchorOffset is 15
PASS textareaAccessible is textareaAccessible.selectionFocusObject
PASS textareaAccessible.selectionFocusOffset is 15
PASS textareaAccessible is textareaAccessible.selectionAnchorObject
PASS textareaAccessible.selectionAnchorOffset is 16
PASS textareaAccessible is textareaAccessible.selectionFocusObject
PASS textareaAccessible.selectionFocusOffset is 16
PASS textareaAccessible is textareaAccessible.selectionAnchorObject
PASS textareaAccessible.selectionAnchorOffset is 17
PASS textareaAccessible is textareaAccessible.selectionFocusObject
PASS textareaAccessible.selectionFocusOffset is 17
PASS textareaAccessible is textareaAccessible.selectionAnchorObject
PASS textareaAccessible.selectionAnchorOffset is 18
PASS textareaAccessible is textareaAccessible.selectionFocusObject
PASS textareaAccessible.selectionFocusOffset is 18
PASS textareaAccessible is textareaAccessible.selectionAnchorObject
PASS textareaAccessible.selectionAnchorOffset is 19
PASS textareaAccessible is textareaAccessible.selectionFocusObject
PASS textareaAccessible.selectionFocusOffset is 19
PASS textareaAccessible is textareaAccessible.selectionAnchorObject
PASS textareaAccessible.selectionAnchorOffset is 20
PASS textareaAccessible is textareaAccessible.selectionFocusObject
PASS textareaAccessible.selectionFocusOffset is 20
PASS emptyTextareaAccessible is textareaAccessible.selectionAnchorObject
PASS textareaAccessible.selectionAnchorOffset is 0
PASS emptyTextareaAccessible is textareaAccessible.selectionFocusObject
PASS textareaAccessible.selectionFocusOffset is 0
PASS textareaAccessible is textareaAccessible.selectionAnchorObject
PASS textareaAccessible.selectionAnchorOffset is 20
PASS textareaAccessible is textareaAccessible.selectionFocusObject
PASS textareaAccessible.selectionFocusOffset is 20
PASS emptyTextareaAccessible is emptyTextareaAccessible.selectionAnchorObject
PASS emptyTextareaAccessible.selectionAnchorOffset is 0
PASS emptyTextareaAccessible is emptyTextareaAccessible.selectionFocusObject
PASS emptyTextareaAccessible.selectionFocusOffset is 0
PASS emptyTextareaAccessible.selectionAnchorOffset is 0
PASS emptyTextareaAccessible.selectionFocusOffset is 0
PASS successfullyParsed is true

TEST COMPLETE

 
